requirements.txt
fire==0.1.3
gast==0.2.0
grpcio==1.14.1
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==2.6.11
mock==3.0.5
numpy==1.14.5
protobuf==3.6.1
regex==2017.4.5
requests==2.21.0
scipy==1.2.1
six==1.11.0
tensorboard==1.13.1
tensorflow==1.12.0
tensorflow-estimator==1.13.0
tensorflow-serving-api==1.12.0
termcolor==1.1.0
tqdm==4.31.1
urllib3==1.24.3
Werkzeug==0.14.1
